+++ tests/integration/components/uic-dropdown-test.js
@@ -6,14 +6,10 @@ moduleForComponent('uic-dropdown', 'Integration | Component | uic dropdown', {
 });
 
 test('it renders', function(assert) {
-  // Set any properties with this.set('myProperty', 'value');
-  // Handle any actions with this.on('myAction', function(val) { ... });
 
   this.render(hbs`{{uic-dropdown}}`);
-
   assert.equal(this.$().text().trim(), '');
 
-  // Template block usage:
   this.render(hbs`
     {{#uic-dropdown}}
       drop down content
@@ -21,7 +17,6 @@ test('it renders', function(assert) {
       template block text
     {{/uic-dropdown}}
   `);
-
   assert.equal(this.$().text().trim(), 'template block text');
 });
 
@@ -51,4 +46,28 @@ test('click', function(assert) {
   this.$('.uic-dropdown button').click();
   assert.equal(this.get('showDropdown'), false);
   assert.equal(this.$('div.dropdown-content').length, 0);
+});
+
+test('click dropdown-container', function(assert) {
+
+  this.set('showDropdown', true);
+  this.set('autoClose', false);
+  this.render(hbs`
+    {{#uic-dropdown showDropdown=showDropdown autoClose=autoClose}}
+      <div class="dropdown-content"></div>
+    {{else}}
+      <button>Click</button>
+    {{/uic-dropdown}}
+    `);
+
+  assert.equal(this.$('.uic-dropdown-container').length, 1);
+  this.$('.uic-dropdown-container').click();
+  assert.equal(this.get('showDropdown'), true);
+  assert.equal(this.$('.uic-dropdown-container').length, 1);
+
+  this.set('autoClose', true);
+
+  this.$('.uic-dropdown-container').click();
+  assert.equal(this.get('showDropdown'), false);
+  assert.equal(this.$('.uic-dropdown-container').length, 0);
 });