+++ CHANGELOG.md
@@ -12,7 +12,26 @@
 
 
 
-### v0.2.0
+
+### v0.4.0 2016-04-28
+
+* [FEATURE] Add `CloseButtonComponent`
+
+* [FEATURE] Add `AutoClose` property to `DropdownComponent` and `DropdownContainerComponent`
+
+
+
+
+### v0.3.0 2016-04-04
+
+* [FEATURE] Add `ButtonGroupComponent`
+
+* [FEATURE] Add `selected` property to `ButtonComponent`
+
+
+
+
+### v0.2.0 2016-03-30
 
 * [FEATURE] Add `ButtonComponent`